Added an excellent test puzzle
hexadoku_sehr_schwer.sudoku is picked up from the german website
http://www.hexadoku.de/
It took about 6 minutes to solve without scanFour, and just under three
minutes with scanFour. I guess it's a keeper.

With this puzzle that takes 3 minutes to solve, it's easy to see if any
future hacks is doing any good.